V 0.7 Released
Configuration Menu
Calibrate spawn point on the detected plane (Click to spawn and then dragged to the desired point)
Room configuration length and width which can be extended to danger or safe zone radius
Save configuration option
Load show with configured parameters
Animation states and names are now sent for audio processing
The absolute location of the phone is being transmitted as well
The absolute position of AR character (wolf) is also being transmitted as well
Distance between phone and Ar character is also being transmitted over the phone.
If AR character is on the boundary so it will come to the tracking area automatically (right now it is abrupt but we can smooth as we liked but it is for just MVP (minimum viable product version))
Using 4 quadrant system considering phone in the centre of the room and according to the position of ARE model it will reposition in a safe area.
During reposition or start of the show, the model will always be facing camera.